# openapi.yml
openapi: 3.0.0
info:
  title: Gironeta Server API
  description: This file defines the communication contract between the AI league application front-end and the back-end service.
  version: 1.0.0
paths:
...
  /api/v0/bot:
    post:
      summary: Create a new bot
      description: This endpoint allows a user to create a new bot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotDTO'
      responses:
        '200':
          description: Bot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotDTO'
        '409':
          description: Conflict, bot already exists
    get:
      summary: Get bots
      description: This endpoint retrieves bots, allowing to filter results by owner.
      parameters:
        - name: owner
          in: query
          description: Filter bots by owner
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of bots retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BotDTO'
  /api/v0/bot/{botId}:
    get:
      summary: Get bot by ID
      description: This endpoint retrieves a bot by its ID.
      parameters:
        - name: botId
          in: path
          required: true
          description: The ID of the bot to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bot retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotDTO'
        '404':
          description: Bot not found
    put:
      summary: Update bot by ID
      description: This endpoint updates a bot by its ID.
      parameters:
        - name: botId
          in: path
          required: true
          description: The ID of the bot to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotDTO'
      responses:
        '200':
          description: Bot updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotDTO'
        '404':
          description: Bot not found
...
  /api/v0/league/{leagueId}/bot:
    post:
      summary: Add bots to a league
      description: This endpoint allows a user to add bots to a league.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to which the bot will be added
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              description: List of bot IDs to add to the league
              items:
                type: integer
                format: int64
      responses:
        '204':
          description: Bots added to league successfully
        '404':
          description: League not found
        '400':
          description: Bad request, invalid bot IDs
...
components:
  schemas:
...
    CreateBotDTO:
      type: object
      properties:
        name:
          type: string
        quality:
          type: string
        imageUrl:
          type: string
        apiUrl:
          type: string
      required:
        - name
        - quality
        - apiUrl
    BotDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        quality:
          type: string
        imageUrl:
          type: string
        apiUrl:
          type: string
        nWins:
          type: integer
        nDraws:
          type: integer
        nLosses:
          type: integer
      required:
        - id
        - name
        - quality
        - apiUrl
...